2/3

+++ tests/test.cpp
@@ -11,11 +11,11 @@ TEST(Parse, Text){
     Json object = Json::parse(json);
     EXPECT_EQ(std::any_cast<std::string>(object["lastname"]), "Ivanov");
     EXPECT_EQ(std::any_cast<bool>(object["islegal"]), false);
-    EXPECT_EQ(std::any_cast<int>(object["age"]), 25);
+    EXPECT_EQ(std::any_cast<double>(object["age"]), 25);
 
     auto marks = std::any_cast<std::vector<std::any>>(object["marks"]);
-    EXPECT_EQ(std::any_cast<int>(marks[0]), 4);
-    EXPECT_EQ(std::any_cast<int>(marks[1]), 5);
+    EXPECT_EQ(std::any_cast<double>(marks[0]), 4);
+    EXPECT_EQ(std::any_cast<double>(marks[1]), 5);
 
     auto address = std::any_cast<std::map<std::string,
     std::any>>(object["address"]);
